// src/interfaces/fetchAllPokemonsResponse.ts
// Generated by https://quicktype.io

export interface FetchAllPokemonsResponse {
  count: number;
  next: null;
  previous: null;
  results: Result[];
}

export interface Result {
  name: string;
  url: string;
}

export interface SimplePokemon {
  id: string;
  name: string;
  // picture:  string;
  // color?: string;
}